TauProducer now inherits from CandidatesProducer
