+++ python/TausProducer.py
@@ -6,7 +6,5 @@ default_configuration = cms.PSet(
         enable = cms.bool(True),
         parameters = cms.PSet(
             src = cms.untracked.InputTag('slimmedTaus'),
-            ea_R03 = cms.untracked.FileInPath('RecoEgamma/ElectronIdentification/data/Summer16/effAreaElectrons_cone03_pfNeuHadronsAndPhotons_80X.txt'),
-            ea_R04 = cms.untracked.FileInPath('cp3_llbb/Framework/data/effAreaElectrons_cone04_pfNeuHadronsAndPhotons.txt'),
             ),
         )
