+++ src/METProducer.cc
@@ -17,11 +17,6 @@ void METProducer::produce(edm::Event& event, const edm::EventSetup& eventSetup) 
     CovMatrix10 = met.getSignificanceMatrix()[1][0];
     CovMatrix11 = met.getSignificanceMatrix()[1][1];
 
-   // std::cout << "...................................................................................." << std::endl;
-   // std::cout << " MET cov reval: " << CovMatrix00 << " " << CovMatrix01 << " " << CovMatrix10 << " " << CovMatrix11 << std::endl;
-   // std::cout << " MET cov nom  : " << CovMatrix00_nom << " " << CovMatrix01_nom << " " << CovMatrix10_nom << " " << CovMatrix11_nom << std::endl;
-  //  std::cout << "...................................................................................." << std::endl;
-
   if (m_slimmed) {
         uncorrectedPt = met.uncorPt();
         uncorrectedPhi = met.uncorPhi();
